Global Variable Management
A global object with the "GameVariableHandler" script can be added to a scene to allow it to modify variables that are used in many other contexts, like an event on the map being able to alter a player's health. A comment in the script itself is provided to (hopefully) explain how to implement it.